testing; bugfixes; dockerization; refactor;

+++ StackAPI.Models/Data/TagDto.cs
@@ -5,6 +5,6 @@ public class TagDto {
   public bool HasSynonyms { get; init; }
   public bool IsModeratorOnly { get; init; }
   public bool IsRequired { get; init; }
-  public decimal Count { get; init; }
+  public long Count { get; init; }
   public string Name { get; init; } = default!;
 }
