+++ StackAPI.Models/Requests/CreateTagRequest.cs
@@ -3,6 +3,6 @@ public class CreateTagRequest {
   public bool HasSynonyms { get; init; }
   public bool IsModeratorOnly { get; init; }
   public bool IsRequired { get; init; }
-  public decimal Count { get; init; }
+  public long Count { get; init; }
   public string Name { get; init; } = default!;
 }
