+++ StackAPI.Models/Responses/TagResponse.cs
@@ -4,6 +4,6 @@ public class TagResponse {
   public bool HasSynonyms { get; init; }
   public bool IsModeratorOnly { get; init; }
   public bool IsRequired { get; init; }
-  public decimal Count { get; init; }
+  public long Count { get; init; }
   public string Name { get; init; } = default!;
 }
